backend/requirements.txt
fastapi
uvicorn
pillow
python-multipart
ultralytics